Incorporacion de fases del boss. v1
Se metio el ataque de caniones, las 3 fases, y un nuevo ataque que hace un area de 270 grados.

-100 neuronas para jony

+++ One'sRedemption/Assets/Boss/Exports/Scripts/Boss/Namespace1.cs
@@ -136,7 +136,7 @@ namespace Utilidades
             }
             else return false;
         }
-        public static bool IsAnimationPlaying(Animator ani, string stateName, int time)
+        public static bool IsAnimationPlaying(Animator ani, string stateName, float time)
         {
             if (ani.GetCurrentAnimatorStateInfo(0).IsName(stateName) && ani.GetCurrentAnimatorStateInfo(0).normalizedTime > time) //si el estado tiene este nombre, y si paso este tiempo (normalizado) devolver verdadero, y viceversa
             {
